Moved NodeSocket class to its own folder, ran tests confirming that it works
