+++ lib/proxy_server.js
@@ -22,7 +22,7 @@ const generateRequests = (testData) => {
     method: testData.tests[0].method,
     maxRequests: parseInt(testData.tests[0].numRequests),
     agentKeepAlive: true,
-    concurrency: 1000,
+    concurrency: 100,
   };
 
   testGenerator.test(config, (error, result) => {
